Move batched mesh manager out
